move pipeline arg generator into its own file
